feat(style): Improve output file naming and markdown style generation
