+++ tests/core/output/styleGenerators/plainStyleGenerator.test.ts
@@ -1,8 +1,8 @@
 import process from 'node:process';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import { buildOutputGeneratorContext } from '../../../src/core/output/outputGenerator.js';
-import { generateXmlStyle } from '../../../src/core/output/xmlStyleGenerator.js';
-import { createMockConfig } from '../../testing/testUtils.js';
+import { buildOutputGeneratorContext } from '../../../../src/core/output/outputGenerator.js';
+import { generatePlainStyle } from '../../../../src/core/output/styleGenerators/plainStyleGenerator.js';
+import { createMockConfig } from '../../../testing/testUtils.js';
 
 vi.mock('fs/promises');
 
@@ -11,11 +11,11 @@ describe('outputGenerator', () => {
     vi.resetAllMocks();
   });
 
-  test('generateXmlOutput should include user-provided header text', async () => {
+  test('generatePlainOutput should include user-provided header text', async () => {
     const mockConfig = createMockConfig({
       output: {
         filePath: 'output.txt',
-        style: 'xml',
+        style: 'plain',
         headerText: 'Custom header text',
         topFilesLength: 2,
         showLineNumbers: false,
@@ -25,11 +25,11 @@ describe('outputGenerator', () => {
     });
 
     const context = await buildOutputGeneratorContext(process.cwd(), mockConfig, [], []);
-    const output = await generateXmlStyle(context);
+    const output = await generatePlainStyle(context);
 
-    expect(output).toContain('file_summary');
-    expect(output).toContain('repository_structure');
+    expect(output).toContain('File Summary');
+    expect(output).toContain('Repository Structure');
     expect(output).toContain('Custom header text');
-    expect(output).toContain('repository_files');
+    expect(output).toContain('Repository Files');
   });
 });
